requirements.txt
gitpython
matplotlib
mesa
networkx==2.2
scipy
seaborn
statsmodels